requirements.txt
#requests
PyYAML
GitPython

pyzmq
